Future proofs order updating by utilising asset types. Expands testing to accomodate. Adds shiny new time type. Expands wrapper websocket functionality definitions

+++ exchanges/kraken/kraken_types.go
@@ -469,10 +469,10 @@ type WsTokenResponse struct {
 }
 
 type wsSystemStatus struct {
-	ConnectionID int64    `json:"connectionID"`
-	Event        string `json:"event"`
-	Status       string `json:"status"`
-	Version      string `json:"version"`
+	ConnectionID float64 `json:"connectionID"`
+	Event        string  `json:"event"`
+	Status       string  `json:"status"`
+	Version      string  `json:"version"`
 }
 
 type wsSubscription struct {
